Fixed Bugs and added back tower preview
I forgot to add the upgrade script back to a few buttons. So they didnt work at first.
And some of the code still made the ui dissapear instead of moving it.

I added back tower preview as it didnt work after i updated it.
It works as it should with less code.